requirements.txt
discord-webhook==0.17.0
websocket-client==1.4.1

websocket~=0.2.1
dateparser~=1.1.1
requests~=2.28.1
emoji~=2.0.0
pandas~=1.4.4
python-telegram-bot==13.14
python-dotenv==0.21.0
